+++ src/core/Config.hpp
@@ -45,6 +45,9 @@ namespace qa::cfg
         // [Speech]
         int focusDedupeMs = 150;
 
+        // [Sounds]
+        int soundVolume = 30;
+
         // [Subtitles]
         bool readSubtitles = true;